CC3XX: Rename driver dir from cc312-rom
Rename the driver to make it clear it is not a ROM-only driver and that
it covers multiple CC3XX family devices. Move it into the generic
arm/drivers directory to make it easier to eventually deprecated the
platform/ext/accelerator directory. Simplify the cmake code for
including the driver, and also split the driver into src and include
directories.

Change-Id: Ie089279f80d67a1ebe17658678ca19928c7bf12a
Signed-off-by: Raef Coles <raef.coles@arm.com>