CC3XX: Allow APIs for configurability of TRNG parameters
Add an API to the CC3XX rng component, called
cc3xx_lowlevel_rng_set_config() that allows to configure
the parameters of the ROSC (ID and subsampling rate, specified
as an inverse in terms of cycles between consecutive samples).
The build time configuration options are used as default if
the API is not called to override them.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I9bab505453602883ecad1b6336747b74b7e948a9