CC3XX: Add recommended_bits_for_generation to curve_data
For curves defined over prime fields, FIPS 186-5 Table A.2
specifies the recommended number of bits to be generated
before a modular reduction operation using the curve order
to generate a non-skewed (unbiased) number to be used as
private key. Other curves have this quantity set to 0 as
this requirement is not available.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I143e87d9b37f83b94b8c71c33acd6dd0570ce82f