RSE: Remove test dependency on CC3XX interface
In the case of RSE, the CC3XX_PLATFORM_INTERFACE CMake variable is set
to platform_bl1_1, because of a circular dependency between
platform_bl1_1 code and the cc3xx low level driver library.

We want to avoid linking the platform_bl1_1 target into the test extras
target as an INTERFACE library as this adds any platform_bl1_1 compiler
flags to the test compile.

Note that the platform_bl1_1 static library itself is still eventually
linked on the command line, as it is still linked to bl1_cc3xx which is
subsequently linked to the cc3xx_test target.

Change-Id: I8201d0ae35815d55d31ec8696d20014ad5d7a51b
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>