CC3XX: Add an helper function in AES for iv_len validation
The PSA layer, or a generic caller, can verify that the length of
the IV being setup is valid for the given mode at a different stage
wrt the actual AES operation setup. In PSA Crypto for example there
is a dedicated API for setting the IV in multipart flows.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I4129e4ddb30acada7f34c63f475ac703860a7ab8